ENH: Add `Q_OBJECT` macro to relevant classes and include `.moc` files
Ensure QObject-derived classes properly use the `Q_OBJECT` macro for
meta-object functionality and include corresponding `.moc` files to
support Qt's meta-object compiler.

Add missing Q_OBJECT to complete Qt inheritance

The Q_OBJECT macro is mandatory for any class that:
   • Inherits from QObject
   • Uses signals or slots
   • Declares properties (Q_PROPERTY)
   • Needs qobject_cast, metaObject(), or other Qt RTTI features

Without it:
   • Qt’s meta-object compiler (moc) won’t generate necessary code
   • You may encounter runtime errors or silently broken signals/slots
   • Clazy and Qt's build system may warn or fail

Need to include full class definitions instead of forward definitions.
Replace forward declarations that were common in Qt 5, but are now
discouraged in Qt 6.

Full declarations are required when the Q_OBJECT macro
is included as recommended to respect the Qt inheritance
expectations.
