+++ Libs/Core/CMakeLists.txt
@@ -71,7 +71,6 @@ set(KIT_SRCS
   ctkHighPrecisionTimer.cpp
   ctkJobScheduler.cpp
   ctkJobScheduler.h
-  ctkJobScheduler_p.h
   ctkLinearValueProxy.cpp
   ctkLinearValueProxy.h
   ctkLogger.cpp
@@ -105,12 +104,38 @@ if(HAVE_BFD)
     )
 endif()
 
+# Headers that should run through moc
+set(KIT_MOC_SRCS
+  ctkAbstractJob.h
+  ctkAbstractWorker.h
+  ctkBooleanMapper.h
+  ctkCallback.h
+  ctkCommandLineParser.h
+  ctkCoreSettings.h
+  ctkErrorLogAbstractMessageHandler.h
+  ctkErrorLogAbstractModel.h
+  ctkErrorLogFDMessageHandler_p.h
+  ctkErrorLogLevel.h
+  ctkErrorLogQtMessageHandler.h
+  ctkErrorLogTerminalOutput.h
+  ctkFileLogger.h
+  ctkJobScheduler.h
+  ctkLinearValueProxy.h
+  ctkLogger.h
+  ctkModelTester.h
+  ctkErrorLogQtMessageHandler.h
+  ctkValueProxy.h
+  ctkWorkflow.h
+  ctkWorkflow_p.h
+  ctkWorkflowStep_p.h
+  ctkWorkflowTransitions.h
+  )
+
 # Abstract class should not be wrapped !
 set_source_files_properties(
   ctkAbstractJob.h
   ctkAbstractWorker.h
   ctkJobScheduler.h
-  ctkJobScheduler_p.h
   WRAP_EXCLUDE
   )
