+++ Libs/Core/ctkErrorLogFDMessageHandler.h
@@ -31,6 +31,7 @@ class ctkErrorLogFDMessageHandlerPrivate;
 /// \ingroup Core
 class CTK_CORE_EXPORT ctkErrorLogFDMessageHandler : public ctkErrorLogAbstractMessageHandler
 {
+  Q_OBJECT
 public:
   typedef ctkErrorLogAbstractMessageHandler Superclass;
