+++ Libs/DICOM/Core/ctkDICOMScheduler.cpp
@@ -23,7 +23,6 @@
 
 // ctkCore includes
 #include <ctkLogger.h>
-#include <ctkJobScheduler_p.h>
 #include <ctkAbstractWorker.h>
 
 // ctkDICOMCore includes
@@ -34,7 +33,6 @@
 #include "ctkDICOMQueryJob.h"
 #include "ctkDICOMRetrieveJob.h"
 #include "ctkDICOMScheduler.h"
-#include "ctkDICOMScheduler_p.h"
 #include "ctkDICOMServer.h"
 #include "ctkDICOMStorageListenerJob.h"
 #include "ctkDICOMUtil.h"
