+++ Libs/DICOM/Widgets/ctkDICOMListenerWidget.h
@@ -29,8 +29,7 @@
 class ctkDICOMListenerWidgetPrivate;
 
 /// \ingroup DICOM_Widgets
-class CTK_DICOM_WIDGETS_EXPORT ctkDICOMListenerWidget : public QWidget
-{
+class CTK_DICOM_WIDGETS_EXPORT ctkDICOMListenerWidget : public QWidget {
   Q_OBJECT
 public:
   typedef QWidget Superclass;
