+++ Libs/DICOM/Widgets/ctkDICOMQueryResultsTabWidget.h
@@ -31,7 +31,7 @@
 /// \ingroup DICOM_Widgets
 class CTK_DICOM_WIDGETS_EXPORT ctkDICOMQueryResultsTabWidget : public QTabWidget
 {
-  Q_OBJECT
+    Q_OBJECT
 public:
   typedef QTabWidget Superclass;
   explicit ctkDICOMQueryResultsTabWidget(QWidget* parent=0);
