+++ Libs/ImageProcessing/ITK/Core/ctkITKErrorLogMessageHandler.h
@@ -35,6 +35,7 @@ class ctkITKErrorLogMessageHandlerPrivate;
 class CTK_IMAGEPROCESSING_ITK_CORE_EXPORT ctkITKErrorLogMessageHandler :
     public ctkErrorLogAbstractMessageHandler
 {
+  Q_OBJECT
 public:
   typedef ctkErrorLogAbstractMessageHandler Superclass;
