+++ Libs/Scripting/Python/Widgets/ctkPythonConsole.cpp
@@ -508,8 +508,8 @@ void ctkPythonConsoleCompleter::updateCompletionModel(const QString& completion)
 // ctkPythonConsolePrivate
 
 //----------------------------------------------------------------------------
-class ctkPythonConsolePrivate : public ctkConsolePrivate
-{
+class ctkPythonConsolePrivate : public ctkConsolePrivate {
+  Q_OBJECT
   Q_DECLARE_PUBLIC(ctkPythonConsole);
 public:
   ctkPythonConsolePrivate(ctkPythonConsole& object);
@@ -728,3 +728,5 @@ void ctkPythonConsole::reset()
 
   this->Superclass::reset();
 }
+
+#include "ctkPythonConsole.moc"
