+++ Libs/Visualization/VTK/Widgets/ctkVTKAbstractMatrixWidget.h
@@ -35,7 +35,7 @@ class ctkVTKAbstractMatrixWidgetPrivate;
 /// of other classes (ctkVTKTransformWidget?) that don't want to expose
 /// setMatrix() publicly.
 class CTK_VISUALIZATION_VTK_WIDGETS_EXPORT ctkVTKAbstractMatrixWidget
-  : public ctkMatrixWidget
+    : public ctkMatrixWidget
 {
   Q_OBJECT
 public:
