+++ Libs/Widgets/ctkCheckableComboBox.cpp
@@ -40,8 +40,8 @@
 #include <ctkCheckableModelHelper.h>
 
 // Similar to QComboBoxDelegate
-class ctkComboBoxDelegate : public QItemDelegate
-{
+class ctkComboBoxDelegate : public QItemDelegate {
+  Q_OBJECT
 public:
     ctkComboBoxDelegate(QObject *parent, QComboBox *cmb)
       : QItemDelegate(parent), ComboBox(cmb)
@@ -432,3 +432,5 @@ void ctkCheckableComboBox::keyReleaseEvent(QKeyEvent *event)
     emit this->returnRelease();
   }
 }
+
+#include "ctkCheckableComboBox.moc"
