+++ Libs/Widgets/ctkCollapsibleGroupBox.cpp
@@ -35,6 +35,7 @@
 //-----------------------------------------------------------------------------
 class ctkCollapsibleGroupBoxStyle:public ctkProxyStyle
 {
+  Q_OBJECT
 public:
   typedef ctkProxyStyle Superclass;
   ctkCollapsibleGroupBoxStyle(QStyle* style = 0, QObject* parent =0)
@@ -330,3 +331,5 @@ bool ctkCollapsibleGroupBox::eventFilter(QObject* child, QEvent* e)
   }
   return this->QGroupBox::eventFilter(child, e);
 }
+
+#include "ctkCollapsibleGroupBox.moc"
