+++ Libs/Widgets/ctkConsole.cpp
@@ -1519,8 +1519,8 @@ QString ctkConsole::stdInRedirectCallBack(void * callData)
 
 namespace
 {
-class InputEventLoop : public QEventLoop
-{
+class InputEventLoop : public QEventLoop {
+  Q_OBJECT
 public:
   InputEventLoop(QApplication * app, QObject * parentObject = 0) :
     QEventLoop(parentObject), App(app){}
@@ -1531,7 +1531,6 @@ public:
   }
   QApplication * App;
 };
-
 }
 
 //-----------------------------------------------------------------------------
@@ -1573,3 +1572,5 @@ void ctkConsole::setMaxVisibleCompleterItems(int count)
   }
   this->completer()->setMaxVisibleItems(count);
 }
+
+#include "ctkConsole.moc"
